called the new methods, moved the new files, and added more test cases
